Updated prog_02.cpp
The logic of file handling through command line arguments are added. The code simply reads the path from the command line argument and creates a text file with a specified name. After creating the text file, the program writes the string from the second command line argument up to x times where "x" is the integer argument specified in the command line.